fixed painting particles, removed DestroyParticle it didn't last long because they changed how this works... yuk
